Implemented --series parameter
Specify a certain series to open a model from via command line parameter --series.

This is useful for preparing shell scripts or desktop shortcuts that specify different ports for a certain series model. Valid arguments are 3, 4, 5, 6 or anything else starting with those numbers, like 6000E or 4000A.
Unrecognized arguments will fall back to the default behaviour of trying all supported series in ascending order, as if the parameter was not given at all.