Corrected scaling for trigger channel
The scale value for the current trigger channel is always the same for the 6000E series (and the same for the external trigger input), but it varies with ADC resolution for other series. On some models it is even different for the external input and the analog channels. This is now taken into account.

Filtered out some illegal ":RANGE" commands that would otherwise spam the Debug Log. They occur if one or more digital channels are added to a WaveformView with analog channels and those voltage range gets adjusted, which does not work for digital channels.